jobs/ingest/requirements.txt
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-api-python-client==2.129.0
google-cloud-storage==2.16.0
google-cloud-firestore==2.16.0
Pillow==10.3.0
pillow-heif==0.16.0
requests==2.32.2
